+++ Neolution.Utilities/Extensions/IConfigurationExtensions.cs
@@ -36,4 +36,19 @@ public static class IConfigurationExtensions
         ArgumentNullException.ThrowIfNull(config);
         return config.GetSection(typeof(T).Name);
     }
+
+    /// <summary>
+    /// Gets the value with the specified key.
+    /// </summary>
+    /// <typeparam name="T">The type of the value.</typeparam>
+    /// <typeparam name="TEnum">The type of the enum.</typeparam>
+    /// <param name="config">The configuration.</param>
+    /// <param name="key">The key.</param>
+    /// <returns>The value.</returns>
+    public static T? GetValue<T, TEnum>(this IConfiguration config, TEnum key)
+        where TEnum : Enum
+    {
+        ArgumentNullException.ThrowIfNull(config);
+        return config.GetValue<T>(key.ToString());
+    }
 }