+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `IServiceCollectionExtensions.AddOptions` with custom section name
+- `IConfigurationExtensions.GetValue` with strongly typed key
 
 ## [2.0.0] - 2026-05-05